backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
minio
python-dotenv
pillow
